Add Update endpoint User/Role

+++ uk223-main/src/main/java/com/example/demo/domain/appUser/UserController.java
@@ -6,17 +6,19 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+
 import javax.management.InstanceAlreadyExistsException;
 import javax.management.InstanceNotFoundException;
 import javax.validation.Valid;
 import java.util.Collection;
 import java.util.UUID;
 
-@RestController @RequestMapping("/api/user")
+@RestController
+@RequestMapping("/api/user")
 @RequiredArgsConstructor
 public class UserController {
-//    ADD YOUR ENDPOINT MAPPINGS HERE
-private final UserService userService;
+    //    ADD YOUR ENDPOINT MAPPINGS HERE
+    private final UserService userService;
 
     @Operation(summary = "Get all Users")
     @PreAuthorize("hasRole('ADMIN')")
@@ -49,24 +51,32 @@ private final UserService userService;
     @Operation(summary = "Delete User by ID")
     @DeleteMapping("/{id}")
     public ResponseEntity<UUID> delete(@PathVariable UUID id) {
-        try {
-            userService.deleteUser(id);
-            return new ResponseEntity<>(id, HttpStatus.OK);
-        } catch (Exception e) {
-            return new ResponseEntity<>(id, HttpStatus.BAD_REQUEST);
+        boolean isUserAuthorized = userService.checkUserAuthorityForEntry(id);
+        if (isUserAuthorized) {
+            try {
+                userService.deleteUser(id);
+                return new ResponseEntity<>(id, HttpStatus.OK);
+            } catch (Exception e) {
+                return new ResponseEntity<>(id, HttpStatus.BAD_REQUEST);
+            }
+        }else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
 
     @Operation(summary = "Update User by ID")
     @PutMapping("/{id}")
     public ResponseEntity<UserUpdateDTO> update(@Valid @RequestBody UserUpdateDTO user, @PathVariable UUID id) {
-        try{
-            return new ResponseEntity<>(userService.update(user, id), HttpStatus.OK);
+        boolean isUserAuthorized = userService.checkUserAuthorityForEntry(id);
+        if (isUserAuthorized) {
+            try {
+                return new ResponseEntity<>(userService.update(user, id), HttpStatus.OK);
+            } catch (Exception e) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
-        catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-
     }
 
 }
